compilation fixes and build script refactors:
- refactor version scripts
- 5.2 compilation fix
- 5.1 compilation fix
- consolidate save asset code
- revised version check
