+++ scripts/prepare_release.bat
@@ -4,12 +4,13 @@
 :: LICENSE file in the root directory of this source tree.
 
 echo off
-IF "%1"=="" GOTO arg_not_exists
-IF "%2"=="" GOTO arg_not_exists
-set version=%1
-set versionName=%2
+
+IF "%1"=="" (GOTO arg_not_exists) else (SET majorVersion=%1)
+IF "%2"=="" (SET minorVersion="0") else (SET minorVersion=%2)
+IF "%3"=="" (SET patchVersion="0") else (SET patchVersion=%3)
+
 node update_API_version.js
-node update_Plugin_version.js %version% %versionName% 
+node update_Plugin_version.js %majorVersion% %minorVersion% %patchVersion%
 goto:eof
 :arg_not_exists
-echo Please run this command with version and versionName. For example: prepare_release.bat 46 46.0.1
+echo Please run this command with version code as 3 arguments for MAJOR MINOR PATCH. For example: prepare_release.bat 46 0 1
