+++ scripts/update_API_version.js
@@ -5,7 +5,7 @@
  * LICENSE file in the root directory of this source tree.
  */
 
-const updateFileAsync = require("./helper");
+const updateFileAsync = require("./updateFileAsync");
 
 const KEY_STRING_START = 'FString ApiVersion{"';
 const KEY_STRING_END = '"';
