+++ scripts/update_Plugin_version.js
@@ -5,7 +5,7 @@
  * LICENSE file in the root directory of this source tree.
  */
 
-const updateFileAsync = require("./helper");
+const updateFileAsync = require("./updateFileAsync");
 
 const VERSION_KEY_STRING_START = '"Version": ';
 const VERSION_KEY_STRING_END = ',';
@@ -14,10 +14,24 @@ const VERSION_NAME_KEY_STRING_END = '"';
 
 console.log(process.argv);
 
-if(process.argv.length < 4) {
-      console.log('Please call this script with : node update_Plugin_version.js 123 1.2.3');
+if(process.argv.length < 3) {
+      console.log('Please call this script with MAJOR MINOR and PATCH version codes : node update_Plugin_version.js 123 0 1');
 }
 
-updateFileAsync('..\\Wit.uplugin', VERSION_KEY_STRING_START,process.argv[2], VERSION_KEY_STRING_END, ()=>{
-  updateFileAsync('..\\Wit.uplugin', VERSION_NAME_KEY_STRING_START,process.argv[3], VERSION_NAME_KEY_STRING_END);
+const majorVersion = process.argv[2];
+let minorVersion = '0'
+let patchVersion = '0'
+
+if(process.argv.length >= 4) {
+  minorVersion = process.argv[3];
+}
+
+if(process.argv.length >= 5) {
+  patchVersion = process.argv[4];
+}
+
+const versionName = majorVersion + '.' + minorVersion + '.' + patchVersion;
+
+updateFileAsync('..\\Wit.uplugin', VERSION_KEY_STRING_START, majorVersion, VERSION_KEY_STRING_END, ()=>{
+  updateFileAsync('..\\Wit.uplugin', VERSION_NAME_KEY_STRING_START, versionName, VERSION_NAME_KEY_STRING_END);
 });
